// konfidence-ui/src/lib/konfidence-api/schema.d.ts
export interface components {
  schemas: {
    ProjectListResponse: {
      data: components["schemas"]["ProjectResponse"][];
    };
    ProjectResponse: {
      id: components["schemas"]["ProjectId"];
      /** @description The name of the project */
      name: string;
    };
    LandscapeListResponse: {
      data: components["schemas"]["LandscapeResponse"][];
    };
    LandscapeResponse: {
      id: components["schemas"]["LandscapeId"];
      /** @description The name of the landscape */
      name: string;
    };
    VectorDeploymentListResponse: {
      data: components["schemas"]["VectorDeploymentResponse"][];
    };
    VectorDeploymentResponse: {
      id: components["schemas"]["VectorDeploymentId"];
      landscapeId: components["schemas"]["LandscapeId"];
      stageId: components["schemas"]["StageId"];
      vector: components["schemas"]["VectorReference"];
      /**
       * @description The vectorDeployment status
       * @enum {string}
       */
      status: "VectorDownloaded" | "ArtifactDeploymentCreated";
    };
    ArtifactDeploymentListResponse: {
      data: components["schemas"]["ArtifactDeploymentResponse"][];
    };
    ArtifactDeploymentResponse: {
      id: components["schemas"]["ArtifactDeploymentId"];
      landscapeId: components["schemas"]["LandscapeId"];
      vectorDeploymentIds: components["schemas"]["VectorDeploymentId"][];
      stageIds: components["schemas"]["StageId"][];
      artifact: components["schemas"]["ArtifactReference"];
      /**
       * @description The artifactDeployment status
       * @enum {string}
       */
      status: "ArtifactFetched" | "ArtifactDeployed";
    };
    StageResponse: {
      id: components["schemas"]["StageId"];
      /** @description The stage name */
      name: string;
      landscapeId: components["schemas"]["LandscapeId"];
      targetStageVersion: components["schemas"]["StageVersion"];
    };
    StageVersion: {
      id: components["schemas"]["StageVersionId"];
      /** @description StageVersion vector */
      vector: components["schemas"]["VectorReference"];
      /** @description Generation of the parent stage */
      stageGeneration: number;
      /** @description Indicates if this is the latest active stageVersion */
      active: boolean;
      /** @enum {string} */
      status: "DeploymentCreated" | "MigrationTasks" | "Active";
    };
    StageListResponse: {
      data: components["schemas"]["StageResponse"][];
    };
    IdentityResponse: {
      name: string;
      givenName: string;
      familyName: string;
      middleName?: string;
      email: string;
      roles: string[];
    };
    ArtifactReference: components["schemas"]["ComponentReference"];
    VectorReference: components["schemas"]["ComponentReference"];
    ComponentReference: {
      repository: string;
      componentName: string;
      componentVersion: string;
    };
    ProjectId: string;
    /** @description The id of a landscape */
    LandscapeId: string;
    /** @description The id of a vectorDeployment */
    VectorDeploymentId: string;
    /** @description The id of an artifactDeployment */
    ArtifactDeploymentId: string;
    /** @description The id of a stage */
    StageId: string;
    StageVersionId: string;
    ErrorResponse: {
      error: {
        code: string;
        message: string;
      };
    };
  };
  responses: {
    /** @description The requested resource was not found. */
    NotFound: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["ErrorResponse"];
      };
    };
    /** @description Authentication required or session expired. */
    Unauthorized: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["ErrorResponse"];
      };
    };
    /** @description Invalid request parameters. */
    BadRequest: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["ErrorResponse"];
      };
    };
    /** @description An unexpected server error occurred. */
    InternalError: {
      headers: {
        [name: string]: unknown;
      };
      content: {
        "application/json": components["schemas"]["ErrorResponse"];
      };
    };
  };
  parameters: {
    /** @description Project Id */
    ProjectPathId: string;
    /** @description Landscape Id */
    LandscapePathId: string;
    /** @description Filter by landscapeId */
    LandscapeQueryId: string;
    /** @description Filter by vectorDeploymentId */
    VectorDeploymentQueryId: string;
  };
  requestBodies: never;
  headers: never;
  pathItems: never;
}